requirements.txt
Django==3.1.4                   # A high-level Python Web framework that encourages rapid development and clean, pragmatic design.
pytest-django==3.8.0            # A Django plugin for pytest.
six==1.15.0                     # Python 2 and 3 compatibility utilities